Adjusted structure for exporters
