+++ src/main/java/pl/bgnat/master/xscrapper/dto/topicmodeling/Document.java
@@ -1,0 +1,11 @@
+package pl.bgnat.master.xscrapper.dto.topicmodeling;
+
+import lombok.Builder;
+
+import java.util.List;
+
+/**
+ * DTO Documentu dla topic modelingu
+ */
+@Builder
+public record Document(String id, String text, List<Long> tweetIds, int tweetsCount) { }
