+++ src/main/java/pl/bgnat/master/xscrapper/utils/CsvWriterUtil.java
@@ -2,11 +2,13 @@ package pl.bgnat.master.xscrapper.utils;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
+import org.springframework.util.StringUtils;
 
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
@@ -14,11 +16,16 @@ import java.time.format.DateTimeFormatter;
 public final class CsvWriterUtil {
     private static final DateTimeFormatter TS = DateTimeFormatter.ofPattern("yyyy-MM-dd_HH-mm-ss");
 
-    /** Zwraca ścieżkę z timestampem, gdy użytkownik nie poda własnej. */
-    public static String defaultName(String prefix, String userPath) {
-        if (userPath != null && !userPath.isBlank()) return userPath;
+    public static String defaultName(String prefix, String userPath, String subfolder) {
         String file = prefix + "_" + LocalDateTime.now().format(TS) + ".csv";
-        return Path.of("").toAbsolutePath().resolve(file).toString();
+        Path dir;
+        if (StringUtils.hasLength(userPath)) {
+            dir = Path.of(userPath, subfolder);
+        } else {
+            dir = Path.of("output/csv", subfolder);
+        }
+        dir.toFile().mkdirs();
+        return dir.resolve(file).toString();
     }
 
     public static FileWriter open(String filePath) throws IOException {
